pyproject.toml

[project]
name = "attn-ensemble-hetero-networks"
version = "0.1.0"
description = "Attention-Driven Ensemble Models for Adaptive Anomaly Detection in Heterogeneous Distributed Networks"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "numpy","scipy","pandas","scikit-learn","torch>=2.0.0","pyyaml","tqdm","matplotlib","rich","pydantic","networkx"
]
[tool.setuptools]
package-dir = {"" = "src"}
[tool.pytest.ini_options]
addopts = "-q"
minversion = "7.0"
